Add trigger for City Beer Store notifications.
Monitor City Beer Store website for changes to beer on tap menu and maintain a list.
When a beer is removed, remove it from the stored list and create an event.
When a beer is added, add it to the stored list and create an event.
When IFTTT hits API point, send events for all matching beer at a the venue for the input Brewery.